+++ src/Facades/InvestorsExchangeData/Deep/TradingStatus.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Digitonic\IexCloudSdk\Facades\InvestorsExchangeData\Deep;
+
+use Illuminate\Support\Facades\Facade;
+
+class TradingStatus extends Facade
+{
+    /**
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return \Digitonic\IexCloudSdk\InvestorsExchangeData\Deep\TradingStatus::class;
+    }
+}