+++ src/InvestorsExchangeData/Stats/Intraday.php
@@ -18,4 +18,12 @@ class Intraday extends BaseGet
     {
         parent::__construct($api);
     }
+
+    /**
+     * @return string
+     */
+    protected function getFullEndpoint(): string
+    {
+        return self::ENDPOINT;
+    }
 }